requirements.txt
appdirs==1.4.4
asgiref==3.3.1
bandit==1.6.2
beautifulsoup4==4.9.3
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.6
django-bootstrap4==2.3.1
django-heroku==0.3.1
django-smartfields==1.1.3
filelock==3.0.12
gitdb==4.0.5
GitPython==3.1.11
gunicorn==20.0.4
install==1.3.4
nose==1.3.7
pbr==5.5.1
psycopg2==2.8.6
pytz==2021.1
PyYAML==5.3.1
six==1.15.0
smmap==3.0.4
soupsieve==2.2
sqlparse==0.4.1
stevedore==3.2.2
virtualenv==20.1.0
whitenoise==5.2.0
